Use separate mapping files for controller and vaults
