+++ harvest-finance/src/vault.ts
@@ -26,19 +26,6 @@ import {
 } from "./harvestUtils";
 import { Address, BigInt, ethereum } from "@graphprotocol/graph-ts";
 
-export function addVaultAndStrategy(call: AddVaultAndStrategyCall): void {
-  getOrCreateHarvestController(call.to.toHexString());
-
-  getOrCreateVault(call.block, call.inputs._vault);
-}
-
-export function setFeeRewardForwarder(call: SetFeeRewardForwarderCall): void {
-  getOrCreateHarvestController(call.to.toHexString());
-
-  let feeRewardForwarder = call.inputs._feeRewardForwarder;
-  getOrCreateFeeRewardForwarder(feeRewardForwarder.toHexString());
-}
-
 /**
  * Handle user deposits to vault
  * @param event
@@ -162,28 +149,6 @@ export function handleWithdraw(event: Withdraw): void {
     rewardTokensBalance,
     null
   );
-}
-
-/**
- * Update vault's input token balances when share price is changed
- * @param event
- */
-export function handleSharePriceChangeLog(event: SharePriceChangeLog): void {
-  getOrCreateHarvestController(event.address.toHexString());
-
-  let market = Market.load(event.params.vault.toHexString()) as Market;
-
-  let vault = getOrCreateVault(event.block, event.params.vault);
-  vault.pricePerShare = event.params.newSharePrice;
-  vault.save();
-
-  let outputTokenBalance = market.outputTokenTotalSupply;
-  let inputTokenBalance = outputTokenBalance.times(vault.pricePerShare).div(vault.underlyingUnit);
-  let inputTokenBalances: TokenBalance[] = [
-    new TokenBalance(vault.underlyingToken, event.params.vault.toHexString(), inputTokenBalance),
-  ];
-
-  updateMarket(event, market, inputTokenBalances, market.outputTokenTotalSupply);
 }
 
 /**
