+++ harvest-finance/subgraph.yaml
@@ -32,7 +32,7 @@ dataSources:
           handler: addVaultAndStrategy
         - function: setFeeRewardForwarder(address)
           handler: setFeeRewardForwarder
-      file: ./src/mapping.ts
+      file: ./src/controller.ts
 
   - kind: ethereum/contract
     name: HarvestEthController2
@@ -64,7 +64,7 @@ dataSources:
           handler: addVaultAndStrategy
         - function: setFeeRewardForwarder(address)
           handler: setFeeRewardForwarder
-      file: ./src/mapping.ts
+      file: ./src/vault.ts
 
 templates:
   - kind: ethereum/contract
@@ -97,7 +97,7 @@ templates:
           handler: handleDeposit
         - event: Withdraw(indexed address,uint256)
           handler: handleWithdraw
-      file: ./src/mapping.ts
+      file: ./src/vault.ts
 
   - kind: ethereum/contract
     name: FeeRewardForwarder
